Add build tests
